+++ api/src/lib.rs
@@ -13,8 +13,6 @@
 )]
 
 pub mod formatter;
-pub mod iri;
-pub mod language_tag;
 pub mod model;
 pub mod parser;